mcpapp/backend/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
email-validator==2.1.0.post1

# 数据库
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# 认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0

# LangChain
langchain==0.0.352
deepseek==0.1.0

# 工具
pydantic==2.5.1
python-dotenv==1.0.0
requests==2.31.0